feat(library): rename KnowledgeLibrary to Library and refactor components into Library/ folder
