+++ src/Ivy.Tendril/Constants.cs
@@ -18,7 +18,7 @@ public static class Constants
     public const int Agent = 80;
     public const int Trash = 90;
     public const int Help = 100;
-    public const int KnowledgeLibrary = 105;
+    public const int Library = 105;
     public const int Onboarding = 110;
 
     public const string DocsUrl = "https://tendril.ivy.app";
